#0: fix to latest main and PR for lamb optim in separate folder   : refactor global sum into upstream
